requirements.txt
openai
python-dotenv
requests
tqdm
pandas